refactor(gateway): split openresponses runtime modules
